+++ iceoryx_posh/test/moduletests/test_popo_publisher_port.cpp
@@ -197,7 +197,7 @@ TEST_F(PublisherPort_test, allocatedChunkContainsPublisherIdAsOriginId)
     auto maybeChunkHeader = m_sutUserSide.allocateChunk(10u);
     auto chunkHeader = maybeChunkHeader.get_value();
 
-    EXPECT_THAT(chunkHeader->m_originId, Eq(static_cast<uint64_t>(m_sutUserSide.getUniqueID())));
+    EXPECT_THAT(chunkHeader->m_originId, Eq(m_sutUserSide.getUniqueID()));
     m_sutUserSide.freeChunk(chunkHeader);
 }
